fix dtype

+++ util/dataset.py
@@ -82,7 +82,7 @@ def get_masks(label, object_name, input_dims= (480, 640), output_dims = (15, 20)
 
         # If there are no objects of interest in the image all the offset get an arbitrary value. 
         # In the loss function all offsets of cell without an object are ignored anyway.
-        offsets = tf.fill((*output_dims, 2), -1)
+        offsets = tf.cast(tf.fill((*output_dims, 2), -1), dtype=tf.float32)
 
         # All cell are marked as false, as there are no object in the whole image.
         objectness_mask = tf.fill(output_dims, False)
